Extract kube TLS name generation function to shareable location
We want tctl to be able to use this function as well, so we're location it
in a place where both tctl and tsh can use it.
Also changing input parameter to k8s host, since it doesn't really care
 about port number.
